chore: move print* functions into common
