+++ common/index.js
@@ -1,4 +1,4 @@
 module.exports = {
-  ...require('./cli-utils'),
-  ...require('./utils'),
-}
+    ...require('./cli-utils'),
+    ...require('./utils'),
+};
