+++ common/utils.js
@@ -1,24 +1,57 @@
-'use strict'
+'use strict';
 
 const { outputJsonSync } = require('fs-extra');
+const chalk = require('chalk');
 
 function loadConfig(env) {
-  return require(`${__dirname}/../axelar-chains-config/info/${env}.json`);
+    return require(`${__dirname}/../axelar-chains-config/info/${env}.json`);
 }
 
 function saveConfig(config, env) {
-  writeJSON(config, `${__dirname}/../axelar-chains-config/info/${env}.json`);
+    writeJSON(config, `${__dirname}/../axelar-chains-config/info/${env}.json`);
 }
 
 const writeJSON = (data, name) => {
-  outputJsonSync(name, data, {
-      spaces: 2,
-      EOL: '\n',
-  });
+    outputJsonSync(name, data, {
+        spaces: 2,
+        EOL: '\n',
+    });
 };
 
-module.exports = {
-  loadConfig,
-  saveConfig,
-  writeJSON,
+const printInfo = (msg, info = '', colour = chalk.green) => {
+    if (info) {
+        console.log(`${msg}: ${colour(info)}\n`);
+    } else {
+        console.log(`${msg}\n`);
+    }
+};
+
+const printWarn = (msg, info = '') => {
+    if (info) {
+        msg = `${msg}: ${info}`;
+    }
+
+    console.log(`${chalk.italic.yellow(msg)}\n`);
+};
+
+const printError = (msg, info = '') => {
+    if (info) {
+        msg = `${msg}: ${info}`;
+    }
+
+    console.log(`${chalk.bold.red(msg)}\n`);
+};
+
+function printLog(log) {
+    console.log(JSON.stringify({ log }, null, 2));
 }
+
+module.exports = {
+    loadConfig,
+    saveConfig,
+    writeJSON,
+    printInfo,
+    printWarn,
+    printError,
+    printLog,
+};
