+++ evm/broadcast-transactions.js
@@ -7,9 +7,8 @@ const {
     providers: { getDefaultProvider },
 } = ethers;
 
-const { printError, printInfo, printWarn, getConfigByChainId, prompt } = require('./utils');
-const { loadConfig } = require('../common');
-const { addBaseOptions } = require('../common');
+const { getConfigByChainId, prompt } = require('./utils');
+const { printError, printInfo, printWarn, addBaseOptions, loadConfig } = require('../common');
 const { sendTransaction, getSignedTx, storeSignedTx } = require('./sign-utils');
 
 async function processCommand(config, _, options, file) {
