+++ evm/check-wallet-balance.js
@@ -4,8 +4,8 @@ const { Command, Option } = require('commander');
 const { ethers } = require('hardhat');
 const { getDefaultProvider, BigNumber } = ethers;
 
-const { printError, mainProcessor } = require('./utils');
-const { addBaseOptions } = require('../common');
+const { mainProcessor } = require('./utils');
+const { printError, addBaseOptions } = require('../common');
 const { getNonceFileData } = require('./sign-utils');
 
 async function processCommand(_, chain, options) {
