+++ evm/deploy-amplifier-gateway.js
@@ -14,10 +14,7 @@ const {
 const {
     saveConfig,
     getBytecodeHash,
-    printInfo,
-    printError,
     printWalletInfo,
-    printWarn,
     prompt,
     mainProcessor,
     deployContract,
@@ -31,7 +28,7 @@ const {
     getDeployOptions,
 } = require('./utils');
 const { calculateDomainSeparator, isValidCosmosAddress } = require('../cosmwasm/utils');
-const { addExtendedOptions } = require('../common');
+const { addExtendedOptions, printInfo, printError, printWarn } = require('../common');
 const { storeSignedTx, signTransaction, getWallet } = require('./sign-utils.js');
 
 const { WEIGHTED_SIGNERS_TYPE, encodeWeightedSigners } = require('@axelar-network/axelar-gmp-sdk-solidity/scripts/utils');
