+++ evm/deploy-contract.js
@@ -9,9 +9,6 @@ const {
 } = ethers;
 const { Command, Option } = require('commander');
 const {
-    printInfo,
-    printWarn,
-    printError,
     getGasOptions,
     isNonEmptyString,
     isNumber,
@@ -27,7 +24,7 @@ const {
     getContractJSON,
     getDeployOptions,
 } = require('./utils');
-const { addExtendedOptions } = require('../common');
+const { addExtendedOptions, printInfo, printWarn, printError } = require('../common');
 
 async function getConstructorArgs(contractName, config, wallet, options) {
     const args = options.args ? JSON.parse(options.args) : {};
