+++ evm/deploy-gateway-v6.2.x.js
@@ -15,13 +15,10 @@ const {
     saveConfig,
     getBytecodeHash,
     verifyContract,
-    printInfo,
     getProxy,
     getEVMAddresses,
     httpGet,
-    printError,
     printWalletInfo,
-    printWarn,
     prompt,
     mainProcessor,
     isContract,
@@ -29,7 +26,7 @@ const {
     getGasOptions,
     getDeployOptions,
 } = require('./utils');
-const { addExtendedOptions } = require('../common');
+const { printInfo, printError, printWarn, addExtendedOptions } = require('../common');
 const { storeSignedTx, signTransaction, getWallet } = require('./sign-utils.js');
 
 const AxelarGatewayProxy = require('@axelar-network/axelar-cgp-solidity/artifacts/contracts/AxelarGatewayProxy.sol/AxelarGatewayProxy.json');
