+++ evm/deploy-its.js
@@ -10,9 +10,6 @@ const {
     deployContract,
     printWalletInfo,
     saveConfig,
-    printInfo,
-    printWarn,
-    printError,
     getContractJSON,
     mainProcessor,
     prompt,
@@ -25,7 +22,7 @@ const {
     getDeployedAddress,
     wasEventEmitted,
 } = require('./utils');
-const { addExtendedOptions } = require('../common');
+const { addExtendedOptions, printInfo, printWarn, printError } = require('../common');
 const { Command, Option } = require('commander');
 
 /**
