+++ evm/deploy-upgradable.js
@@ -12,17 +12,8 @@ const IUpgradable = require('@axelar-network/axelar-gmp-sdk-solidity/interfaces/
 const { Command, Option } = require('commander');
 
 const { deployUpgradable, deployCreate2Upgradable, deployCreate3Upgradable, upgradeUpgradable } = require('./upgradable');
-const {
-    printInfo,
-    printError,
-    printWalletInfo,
-    getDeployedAddress,
-    prompt,
-    getGasOptions,
-    getDeployOptions,
-    mainProcessor,
-} = require('./utils');
-const { addExtendedOptions } = require('../common');
+const { printWalletInfo, getDeployedAddress, prompt, getGasOptions, getDeployOptions, mainProcessor } = require('./utils');
+const { addExtendedOptions, printInfo, printError } = require('../common');
 
 function getProxy(wallet, proxyAddress) {
     return new Contract(proxyAddress, IUpgradable.abi, wallet);
