+++ evm/governance.js
@@ -10,7 +10,6 @@ const {
 } = ethers;
 const { Command, Option } = require('commander');
 const {
-    printInfo,
     getGasOptions,
     printWalletInfo,
     isValidTimeFormat,
@@ -18,7 +17,6 @@ const {
     etaToDate,
     getCurrentTimeInSeconds,
     wasEventEmitted,
-    printWarn,
     getBytecodeHash,
     isValidAddress,
     mainProcessor,
@@ -28,6 +26,7 @@ const {
     writeJSON,
     isKeccak256Hash,
 } = require('./utils.js');
+const { printInfo, printWarn } = require('../common.js');
 const { addBaseOptions } = require('./cli-utils.js');
 const { getWallet } = require('./sign-utils.js');
 const IAxelarServiceGovernance = require('@axelar-network/axelar-gmp-sdk-solidity/interfaces/IAxelarServiceGovernance.json');
