+++ evm/verify-contract.js
@@ -9,8 +9,9 @@ const {
     Contract,
 } = ethers;
 const { Command, Option } = require('commander');
-const { verifyContract, getEVMAddresses, printInfo, printError, mainProcessor, getContractJSON, validateParameters } = require('./utils');
-const { addBaseOptions } = require('../common');
+const { verifyContract, getEVMAddresses, mainProcessor, getContractJSON, validateParameters } = require('./utils');
+
+const { printInfo, printError, addBaseOptions } = require('../common');
 const { getTrustedChainsAndAddresses } = require('./its');
 
 async function processCommand(config, chain, options) {
